+++ ch_frb_rfi/__init__.py
@@ -17,7 +17,7 @@ For a list of all acquisitions available, see dir(ch_frb_rfi.acquisitions).
 """
 
 # These functions are imported to the top level of the ch_frb_rfi package.
-from chains import clipper_chain, \
+from .chains import clipper_chain, \
     detrender_chain, \
     transform_chain, \
     transform_parameters